+++ IsekaiMod/Content/Features/IsekaiProtagonist/Archetypes/Hero/TrueMark.cs
@@ -84,7 +84,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.Hero
                     c.Actions = ActionFlow.DoSingle<ContextActionApplyBuff>(c => {
                         c.Permanent = true;
                         c.m_Buff = TrueMarkBuff.ToReference<BlueprintBuffReference>();
-                        c.DurationValue = Constants.ZeroDuration;
+                        c.DurationValue = Constants.Duration.Zero;
                         c.IsNotDispelable = true;
                     });
                 });