+++ IsekaiMod/Content/Features/IsekaiProtagonist/SpecialPower/SummonBeast.cs
@@ -155,7 +155,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.SpecialPower
                 AfterSpawn = ActionFlow.DoSingle<ContextActionApplyBuff>(c => {
                     c.Permanent = true;
                     c.m_Buff = SummonedCreatureSpawnMonsterVI_IX.ToReference<BlueprintBuffReference>();
-                    c.DurationValue = Constants.ZeroDuration;
+                    c.DurationValue = Constants.Duration.Zero;
                     c.IsNotDispelable = true;
                 }),
             };