+++ IsekaiMod/Content/Features/IsekaiProtagonist/SummonHarem.cs
@@ -114,7 +114,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist
                 AfterSpawn = ActionFlow.DoSingle<ContextActionApplyBuff>(c => {
                     c.Permanent = true;
                     c.m_Buff = SummonedCreatureSpawnMonsterVI_IX.ToReference<BlueprintBuffReference>();
-                    c.DurationValue = Constants.ZeroDuration;
+                    c.DurationValue = Constants.Duration.Zero;
                     c.IsNotDispelable = true;
                 }),
             };